backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
fastapi==0.131.0
h11==0.16.0
idna==3.11
ImageIO==2.37.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lazy_loader==0.4
llvmlite==0.46.0
networkx==3.6.1
numba==0.64.0
numpy==2.4.2
packaging==26.0
pillow==12.1.1
platformdirs==4.9.2
pooch==1.9.0
pydantic==2.12.5
pydantic_core==2.41.5
PyMatting==1.1.15
python-multipart==0.0.22
referencing==0.37.0
rembg==2.0.72
requests==2.32.5
rpds-py==0.30.0
scikit-image==0.26.0
scipy==1.17.1
starlette==0.52.1
tifffile==2026.2.20
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
